get Single Book endpoint added

+++ api/routes/books.py
@@ -1,6 +1,6 @@
 from typing import OrderedDict
 
-from fastapi import APIRouter, status
+from fastapi import APIRouter, status, HTTPException
 from fastapi.responses import JSONResponse
 
 from api.db.schemas import Book, Genre, InMemoryDB
@@ -50,6 +50,8 @@ async def get_books() -> OrderedDict[int, Book]:
 
 @router.put("/{book_id}", response_model=Book, status_code=status.HTTP_200_OK)
 async def update_book(book_id: int, book: Book) -> Book:
+    if book_id not in db.books:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Book not found")
     return JSONResponse(
         status_code=status.HTTP_200_OK,
         content=db.update_book(book_id, book).model_dump(),
@@ -58,5 +60,14 @@ async def update_book(book_id: int, book: Book) -> Book:
 
 @router.delete("/{book_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_book(book_id: int) -> None:
+    if book_id not in db.books:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Book not found")
     db.delete_book(book_id)
     return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=None)
+
+
+@router.get("/{book_id}", response_model= Book,  status_code=status.HTTP_200_OK)
+async def get_book(book_id: int) -> Book:
+    if book_id not in db.books:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Book not found")
+    return db.get_book(book_id)
